Added Xcode 8.3.1 headers.

+++ DVTFoundation/DVTDiffDescriptor.h
@@ -24,8 +24,8 @@
     DVTDiffDescriptor *_parentDiffDescriptor;
     NSMutableArray *_childDescriptors;
     NSDictionary *_subdescriptorResults;
-    int _diffDescriptorType;
-    int _diffMergeDirection;
+    long long _diffDescriptorType;
+    long long _diffMergeDirection;
     BOOL _resolved;
     BOOL _binaryConflict;
     BOOL _diffMergeDirectionLocked;
@@ -41,7 +41,7 @@
 @property struct _NSRange originalContentRange; // @synthesize originalContentRange=_originalContentRange;
 @property unsigned long long modifiedTokenHash; // @synthesize modifiedTokenHash=_modifiedTokenHash;
 @property unsigned long long originalTokenHash; // @synthesize originalTokenHash=_originalTokenHash;
-@property int diffDescriptorType; // @synthesize diffDescriptorType=_diffDescriptorType;
+@property long long diffDescriptorType; // @synthesize diffDescriptorType=_diffDescriptorType;
 @property struct _NSRange modifiedTokenRange; // @synthesize modifiedTokenRange=_modifiedTokenRange;
 @property struct _NSRange originalTokenRange; // @synthesize originalTokenRange=_originalTokenRange;
 @property DVTDiffContext *diffContext; // @synthesize diffContext=_diffContext;
@@ -59,7 +59,7 @@
 - (void)insertObject:(id)arg1 inChildDiffDescriptorsAtIndex:(long long)arg2;
 - (void)_diffDescriptor:(id)arg1 willRemoveChildDiffDescriptor:(id)arg2;
 - (void)_diffDescriptor:(id)arg1 didAddChildDiffDescriptor:(id)arg2;
-@property int diffMergeDirection; // @dynamic diffMergeDirection;
+@property long long diffMergeDirection; // @dynamic diffMergeDirection;
 - (id)childDiffDescriptors;
 - (unsigned long long)dvt_diffHashForDataSource:(id)arg1;
 - (id)description;
