+++ IDEiOSSupportCore/IDEBuildable-Protocol.h
@@ -19,7 +19,7 @@
 @property(readonly, copy) IDEEntityIdentifier *entityIdentifier;
 @property(readonly, copy) NSString *buildableIdentifier;
 @property(readonly, copy) NSString *productTypeIdentifier;
-- (IDEBuilder *)createBuilderForBuildCommand:(int)arg1 withBuildTaskQueueSet:(IDEBuildOperationQueueSet *)arg2 parameters:(IDEBuildParameters *)arg3 buildOnlyTheseFiles:(NSArray *)arg4 restorePersistedBuildResults:(BOOL)arg5 withManager:(id <IDEBuildableOperationManager>)arg6;
+- (IDEBuilder *)createBuilderForBuildCommand:(long long)arg1 withBuildTaskQueueSet:(IDEBuildOperationQueueSet *)arg2 parameters:(IDEBuildParameters *)arg3 buildOnlyTheseFiles:(NSArray *)arg4 restorePersistedBuildResults:(BOOL)arg5 withManager:(id <IDEBuildableOperationManager>)arg6;
 - (NSArray *)implicitDependenciesForBuildParameters:(IDEBuildParameters *)arg1 executionEnvironment:(IDEExecutionEnvironment *)arg2 returningMessages:(id *)arg3;
 - (NSArray *)directDependencies;
 - (NSArray *)orderedRecursiveDependenciesIncludingSelf:(BOOL)arg1;
